posthog analytics
- Added new component based posthog event with parameters to fire better events.
